Allow adding rules in batches
The rules API is rate limited to 5 requests every 5 seconds and there
are situations where a large number of rules need to be modified beyond
this limit(E.g. during first application startup)

This is a minor change that adds an addRules and deleteRules method to
the GnipFacade that takes a Rules object to facilitate adding/deleting
rules in batches.